fix: bug and refactoring
