+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mochi-inc-japan/algolia-tools",
+  "name": "@mochi-inc-japan/algolia-firebase-tools",
   "version": "1.1.1",
   "main": "dist/index.js",
   "bin": {
@@ -52,7 +52,8 @@
   },
   "peerDependencies": {
     "algoliasearch": "^4.3.0",
-    "typescript": "^3.9.6"
+    "typescript": "^3.9.6",
+    "firebase-admin": "^9.0.0"
   },
   "husky": {
     "hooks": {
