+++ src/plugin/FirebaseInvoke.ts
@@ -1,21 +1,21 @@
 import * as firebaseAdmin from 'firebase-admin'
 import { AlgoliaIndexManager } from '../util/AlgoliaIndexManager'
 
-export type FirebaseInvoke = {
+export type FirebaseInvokeInternal = {
   admin: typeof firebaseAdmin
   algoliaManager: AlgoliaIndexManager
   batchTimeKey?: string
 }
 
-export default class AlgoliaFirebaseManager {
-  public constructor(args: FirebaseInvoke) {
+export default class FirebaseInvokeClass {
+  public constructor(args: FirebaseInvokeInternal) {
     this.admin = args.admin
     this.algoliaManager = args.algoliaManager
     this.batchTimeKey = args.batchTimeKey || 'algolia-send-index-batchtime'
   }
-  public batchTimeKey: FirebaseInvoke['batchTimeKey']
-  public admin: FirebaseInvoke['admin']
-  public algoliaManager: FirebaseInvoke['algoliaManager']
+  public batchTimeKey: FirebaseInvokeInternal['batchTimeKey']
+  public admin: FirebaseInvokeInternal['admin']
+  public algoliaManager: FirebaseInvokeInternal['algoliaManager']
 
   public batchSendDataToIndex = async (
     {
@@ -85,6 +85,30 @@ export default class AlgoliaFirebaseManager {
     } catch (e) {
       console.error(e)
       return false
+    }
+  }
+
+  public async removeAllDataFromIndex(indexName: string[]) {
+    try {
+      await Promise.all(
+        indexName.map((_indexName: string) =>
+          this.algoliaManager.removeAllDataFromIndex(_indexName)
+        )
+      )
+    } catch (e) {
+      console.log('deleteIndex was failed.')
+    }
+    try {
+      await Promise.all(
+        indexName.map((_indexName: string) =>
+          this.admin
+            .database()
+            .ref(`${this.batchTimeKey}/${_indexName}`)
+            .remove()
+        )
+      )
+    } catch (e) {
+      console.log('deleteIndex was failed.')
     }
   }
 }
