+++ template/const.ts
@@ -1,0 +1,4 @@
+export const ALGOLIA_ID = process.env.ALGOLIA_ID || ''
+export const ALGOLIA_ADMIN_KEY = process.env.ALGOLIA_ADMIN_KEY || ''
+export const ALGOLIA_SEARCH_KEY = process.env.ALGOLIA_SEARCH_KEY
+export const INDEX_NAMESPACE = process.env.INDEX_NAMESPACE
